+++ public/admin/_app/immutable/nodes/14.BShKg6uR.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/CbB0GLto2.js";export{e as component};