+++ public/admin/_app/immutable/nodes/24.Q2tDMg4V.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/CLlKAAuU.js";export{e as component};